rename: features/detail-page/FloatingBar.tsx로 수정
