+++ app/page.tsx
@@ -1,5 +1,5 @@
 /* eslint-disable import/extensions */
-import FloatingBar from "features/FloatingBar";
+import FloatingBar from "features/detail-page/FloatingBar";
 
 export default function Home() {
   return <FloatingBar />;
